backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
motor==3.3.2
pydantic==2.5.0
starlette==0.27.0
python-telegram-bot==20.7
google-auth==2.23.4
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
httpx==0.25.2
asyncio==3.4.3
aiofiles==23.2.1